v0.2 cont: UX polish, skill tree refund, decimal regen, menu music, chore: remove Zap VFX demo assets
- SkillTreeManager: refund ability point (unspend last node)
- HealthSystem: decimal HP/regen values, float fields throughout
- ExperienceManager: level-up VFX animation hook
- MusicManager: new file — stop menu music on GameScene load
- CursorManager: fix flicker when plugging in controller
- EquipSlotUI: new file — highlight equipment slot on item hover
- DraggableItem / InventoryUI: drag interaction fixes from v0.2 test
- MainMenuManager + MainMenu.uxml: inspector-configurable UI labels
- DevLog, EnemyAI, EnemyHealthBar, PlayerController, WaveManager, DeathScreenManager: minor fixes
- chore: remove Assets/FX/Vefects/Zap VFX URP demo content (unused sample)
- docs: roadmap update

+++ Assets/Scripts/EnemyAI.cs
@@ -50,11 +50,11 @@ public class EnemyAI : NetworkBehaviour
     // Read by EnemyHealthBar on all clients to drive the HP bar UI.
 
     [HideInInspector]
-    public NetworkVariable<int> NetworkHealth = new NetworkVariable<int>(
+    public NetworkVariable<float> NetworkHealth = new NetworkVariable<float>(
         0, NetworkVariableReadPermission.Everyone, NetworkVariableWritePermission.Server);
 
     [HideInInspector]
-    public NetworkVariable<int> NetworkMaxHealth = new NetworkVariable<int>(
+    public NetworkVariable<float> NetworkMaxHealth = new NetworkVariable<float>(
         0, NetworkVariableReadPermission.Everyone, NetworkVariableWritePermission.Server);
 
     // ─── State ────────────────────────────────────────────────────────────────
